requirements.txt
asgiref==3.7.2
Django==4.2.2
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.13
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
